:fire: remove borax.dic module
