--- docs/guides/alias_dictionary.md
@@ -1,7 +1,0 @@
-# AliasDictionary 模块
-
-> 模块： `borax.structures.dictionary`
-
-## 概述
-
-别名字典(`AliasDictionary`)， 允许定义一个字典的键的别名，并使用该别名访问字典的内容。
